Several fixes to get code golint and go vet clean
